refactor: 타입 안전성, 렌더링 성능, 번들 크기 개선
