+++ mission/chapter07/src/pages/users/UsersPage.tsx
@@ -1,213 +1,8 @@
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { useRef, useState } from 'react';
+import { useQuery } from '@tanstack/react-query';
+import { useState } from 'react';
 import { useParams } from 'react-router';
-import { uploadImage } from '../../apis/uploadsApi';
-import { getMyInfo, getUserInfo, updateMyInfo } from '../../apis/usersApi';
-import { useAuth } from '../../contexts/AuthContext';
-import type { UserInfo } from '../../types/user';
-
-function ProfileEditModal({
-  initialName,
-  initialBio,
-  initialAvatar,
-  onClose,
-}: {
-  initialName: string;
-  initialBio: string;
-  initialAvatar: string;
-  onClose: () => void;
-}) {
-  const queryClient = useQueryClient();
-  const { userName, updateUserName } = useAuth();
-  const overlayRef = useRef<HTMLDivElement>(null);
-
-  const [name, setName] = useState(initialName);
-  const [bio, setBio] = useState(initialBio);
-  const [avatarFile, setAvatarFile] = useState<File | null>(null);
-  const [avatarPreview, setAvatarPreview] = useState<string>(initialAvatar);
-  const [error, setError] = useState('');
-
-  const updateMutation = useMutation({
-    mutationFn: async () => {
-      let avatarUrl: string | null = avatarPreview || null;
-      if (avatarFile) {
-        try {
-          avatarUrl = await uploadImage(avatarFile);
-        } catch {
-          // 업로드 실패 시 기존 avatar 유지
-        }
-      }
-      return updateMyInfo({
-        name: name.trim() || undefined,
-        bio: bio.trim() || null,
-        avatar: avatarUrl,
-      });
-    },
-
-    // ── 낙관적 업데이트 ──────────────────────────────────────
-    onMutate: async () => {
-      // 1. 진행 중인 리페치 취소 (덮어쓰기 방지)
-      await queryClient.cancelQueries({ queryKey: ['user', 'me'] });
-      await queryClient.cancelQueries({ queryKey: ['myInfo'] });
-
-      // 2. 이전 데이터 저장 (롤백용)
-      const previousUser = queryClient.getQueryData<UserInfo>(['user', 'me']);
-      const previousMyInfo = queryClient.getQueryData<UserInfo>(['myInfo']);
-      const previousUserName = userName;
-
-      // 3. 캐시 낙관적 업데이트
-      const optimisticUpdate = (old: UserInfo | undefined) =>
-        old ? { ...old, name: name.trim(), bio: bio.trim() || null } : old;
-      queryClient.setQueryData<UserInfo>(['user', 'me'], optimisticUpdate);
-      queryClient.setQueryData<UserInfo>(['myInfo'], optimisticUpdate);
-
-      // 4. Nav-Bar 닉네임 즉시 업데이트
-      updateUserName(name.trim());
-
-      return { previousUser, previousMyInfo, previousUserName };
-    },
-
-    // ── 실패 시 롤백 ──────────────────────────────────────────
-    onError: (err, _, context) => {
-      if (context?.previousUser !== undefined) {
-        queryClient.setQueryData(['user', 'me'], context.previousUser);
-      }
-      if (context?.previousMyInfo !== undefined) {
-        queryClient.setQueryData(['myInfo'], context.previousMyInfo);
-      }
-      // Nav-Bar 닉네임 롤백
-      if (context?.previousUserName !== undefined) {
-        updateUserName(context.previousUserName ?? '');
-      }
-      setError(err instanceof Error ? err.message : '프로필 수정에 실패했습니다.');
-    },
-
-    // ── 성공 시 서버 응답으로 최종 동기화 ─────────────────────
-    onSuccess: (data) => {
-      if (data?.name) updateUserName(data.name);
-      onClose();
-    },
-
-    // ── 항상 최종 서버 상태와 동기화 ───────────────────────────
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ['user', 'me'] });
-      queryClient.invalidateQueries({ queryKey: ['myInfo'] });
-    },
-  });
-
-  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    setAvatarFile(file);
-    const reader = new FileReader();
-    reader.onload = () => setAvatarPreview(reader.result as string);
-    reader.readAsDataURL(file);
-  };
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    setError('');
-    if (!name.trim()) {
-      setError('이름을 입력해주세요.');
-      return;
-    }
-    updateMutation.mutate();
-  };
-
-  const handleOverlayClick = (e: React.MouseEvent) => {
-    if (e.target === overlayRef.current) onClose();
-  };
-
-  return (
-    <div
-      ref={overlayRef}
-      className="fixed inset-0 z-50 flex items-center justify-center bg-black/70 px-4"
-      onClick={handleOverlayClick}
-    >
-      <div className="w-full max-w-md rounded-2xl bg-[#1e1e1e] border border-white/10 shadow-2xl">
-        <div className="flex items-center justify-between border-b border-white/10 px-6 py-4">
-          <h2 className="text-lg font-bold text-white">프로필 수정</h2>
-          <button
-            onClick={onClose}
-            className="rounded-md p-1 text-slate-400 hover:text-white transition-colors"
-            aria-label="닫기"
-          >
-            <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-              <path d="M18 6 6 18M6 6l12 12" />
-            </svg>
-          </button>
-        </div>
-
-        <form onSubmit={handleSubmit} className="flex flex-col gap-5 p-6">
-          {/* 프로필 사진 */}
-          <div className="flex flex-col items-center gap-3">
-            <label className="cursor-pointer">
-              {avatarPreview ? (
-                <img
-                  src={avatarPreview}
-                  alt="프로필 사진"
-                  className="h-20 w-20 rounded-full object-cover ring-2 ring-pink-500/50"
-                />
-              ) : (
-                <div className="flex h-20 w-20 items-center justify-center rounded-full bg-pink-500/20 text-2xl font-bold text-pink-400">
-                  {initialName.charAt(0).toUpperCase()}
-                </div>
-              )}
-              <input
-                type="file"
-                accept="image/*"
-                className="hidden"
-                onChange={handleFileChange}
-              />
-            </label>
-            <span className="text-xs text-slate-500">클릭하여 사진 변경 (선택)</span>
-          </div>
-
-          {/* 이름 */}
-          <div>
-            <label htmlFor="profile-name" className="mb-1.5 block text-sm font-medium text-slate-300">
-              이름 <span className="text-pink-500">*</span>
-            </label>
-            <input
-              id="profile-name"
-              value={name}
-              onChange={(e) => setName(e.target.value)}
-              placeholder="이름을 입력하세요"
-              className="w-full rounded-lg border border-white/10 bg-white/5 px-3 py-2 text-sm text-white placeholder-slate-500 outline-none transition-colors focus:border-pink-500/50"
-            />
-          </div>
-
-          {/* Bio */}
-          <div>
-            <label htmlFor="profile-bio" className="mb-1.5 block text-sm font-medium text-slate-300">
-              자기소개 <span className="text-slate-500">(선택)</span>
-            </label>
-            <textarea
-              id="profile-bio"
-              value={bio}
-              onChange={(e) => setBio(e.target.value)}
-              placeholder="자기소개를 입력하세요"
-              rows={3}
-              className="w-full resize-none rounded-lg border border-white/10 bg-white/5 px-3 py-2 text-sm text-white placeholder-slate-500 outline-none transition-colors focus:border-pink-500/50"
-            />
-          </div>
-
-          {error && (
-            <p className="rounded-lg bg-red-500/10 px-3 py-2 text-sm text-red-400">{error}</p>
-          )}
-
-          <button
-            type="submit"
-            disabled={updateMutation.isPending}
-            className="h-11 w-full rounded-xl bg-pink-500 text-sm font-semibold text-white hover:bg-pink-600 transition-colors disabled:cursor-not-allowed disabled:opacity-60"
-          >
-            {updateMutation.isPending ? '저장 중...' : '저장하기'}
-          </button>
-        </form>
-      </div>
-    </div>
-  );
-}
+import { getMyInfo, getUserInfo } from '../../apis/usersApi';
+import ProfileEditModal from '../../components/modals/ProfileEditModal';
 
 function UsersPage() {
   const { userId } = useParams();
